+++ src/model.rs
@@ -458,6 +458,7 @@ pub struct Stats {
     pub cuts: i32,
     pub capture_share: i32,
     pub nodes_per_ms: i32,
+    pub logging: Vec<String>,
 }
 
 impl Stats {
@@ -474,13 +475,19 @@ impl Stats {
             cuts: 0,
             capture_share: 0,
             nodes_per_ms: 0,
+            logging: Vec::default(),
          }
+    }
+
+    pub fn add_log(&mut self, msg: String) {
+        self.logging.push(msg.to_string());
     }
 
     pub fn calculate(&mut self) {
         self.cuts = 100 - (self.calculated_nodes as i32 * 100 / self.created_nodes as i32);
         self.capture_share = self.created_capture_node as i32 * 100 / self.created_nodes as i32;
         self.nodes_per_ms = self.created_nodes as i32 / self.calc_time_ms as i32 + 1;
+        self.zobrist_hit = self.zobrist_hit * 100 / self.eval_nodes;
     }
 
     pub fn add_created_nodes(&mut self, value: usize) {
@@ -521,6 +528,7 @@ impl Stats {
         self.eval_nodes = 0;
         self.calc_time_ms = 0;
         self.zobrist_hit = 0;
+        self.logging = Vec::default();
     }
 
     pub fn to_string(&self) -> String {