+++ src/notation_util.rs
@@ -1,5 +1,3 @@
-use std::collections::VecDeque;
-
 use crate::model::Turn;
 use regex::Regex;